fix: address 360-degree audit quick fixes
- Fix ADR format inconsistencies (008, 011, 012, 020) — standardize
  Status/Date format, add missing Decision and Alternatives sections
- Fix stale ADR count in README project structure
- Add expected outcomes to README Usage section
- Fix homepage genre list (5 → 9 genres)
- Move GenreLens type from component to src/types/ (fix dep direction)

Closes #501, closes #502, closes #503, closes #508, closes #512
