+++ src/components/interactive/GenreGuide/types.ts
@@ -1,34 +1,5 @@
 import type { Genre } from "../../../types/genre";
-import type { Brand, Mount } from "../../../types/common";
-
-interface GenreLens {
-  brand: Brand;
-  model: string;
-  mount: Mount;
-  type: "prime" | "zoom";
-  focalLengthMin: number;
-  focalLengthMax: number;
-  maxAperture: number;
-  sweetSpotAperture?: number;
-  maxMagnification?: number;
-  hasOis?: boolean;
-  isWeatherSealed?: boolean;
-  isDiscontinued?: boolean;
-  weight: number;
-  price: number;
-  genreMarks?: Partial<Record<Genre, number>>;
-  editorialPicks?: Genre[];
-  centerStopped?: number;
-  cornerStopped?: number;
-  centerWideOpen?: number;
-  astigmatism?: number;
-  coma?: number;
-  longitudinalCA?: number;
-  lateralCA?: number;
-  distortion?: number;
-  bokeh?: number;
-  flareResistance?: number;
-}
+import type { GenreLens } from "../../../types/genre-lens";
 
 interface GenreGuideProps {
   lenses: GenreLens[];
