+++ src/pages/index.astro
@@ -84,7 +84,7 @@ const wikiCount = wikiEntries.length;
       <ol class="workflow-steps">
         <li>
           <a href="/genre">Pick a genre</a> — nightscape, landscape, portrait, street,
-          or architecture
+          architecture, travel, sport, wildlife, or macro
         </li>
         <li>Select the scene you want to shoot and check the conditions</li>
         <li>See which lenses score best and decide what to rent or buy</li>
